+++ packages/terra-doc-template/examples/test-examples/DefaultDocTemplate.example.jsx
@@ -10,8 +10,8 @@ import testComponentSrc from '!raw-loader!./TestComponent';
 
 const Index = () => {
   const propsTables = [
-    { componentSource: testComponentSrc, componentName: 'Test Component' },
-    { componentSource: testComponentSrc, componentName: 'Test Component (Again)' },
+    { componentSrc: testComponentSrc, componentName: 'Test Component' },
+    { componentSrc: testComponentSrc, componentName: 'Test Component (Again)' },
   ];
   const exampleElement = <TestComponentExample />;
   const examples = [
@@ -21,7 +21,13 @@ const Index = () => {
 
   return (
     <div id="DocTemplateContainer">
-      <DocTemplate packageName="test-package-name" readme={readme} propsTables={propsTables} examples={examples} />
+      <DocTemplate
+        packageName="test-package-name"
+        readme={readme}
+        srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-doc-template"
+        propsTables={propsTables}
+        examples={examples}
+      />
     </div>
   );
 };