+++ packages/terra-doc-template/src/DocTemplate.jsx
@@ -10,13 +10,17 @@ const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
-   * The given component's npm package name
+   * The given component's npm package name.
    */
   packageName: PropTypes.string,
   /**
-   * The given component's readme file imported to a string
+   * The given component's readme file imported to a string.
    */
   readme: PropTypes.string,
+  /**
+   * The url to the source code for the given component.
+   */
+  srcPath: PropTypes.string,
   /**
    * All of the example(s) that will be displayed. An empty array is supported.
    * ```
@@ -56,11 +60,12 @@ const propTypes = {
 const defaultProps = {
   packageName: '',
   readme: '',
+  srcPath: '',
   examples: [],
   propsTables: [],
 };
 
-const DocTemplate = ({ packageName, readme, examples, propsTables, ...customProps }) => {
+const DocTemplate = ({ packageName, readme, srcPath, examples, propsTables, ...customProps }) => {
   let id = 0;
   const localExamples = examples;
   const localPropsTables = propsTables;
@@ -82,6 +87,7 @@ const DocTemplate = ({ packageName, readme, examples, propsTables, ...customProp
     <div {...customProps}>
       {packageName && <Markdown src={version} />}
       {readme && <Markdown src={readme} />}
+      {srcPath && <a href={srcPath}>View component source code</a>}
 
       {localExamples.length > 0 && <h1 className={cx('.examples-header')} >Examples</h1>}
       {localExamples.map(example =>
@@ -96,7 +102,7 @@ const DocTemplate = ({ packageName, readme, examples, propsTables, ...customProp
 
       {localPropsTables.map(propsTable =>
         <PropsTable
-          src={propsTable.componentSource}
+          src={propsTable.componentSrc}
           componentName={propsTable.componentName}
           key={propsTable.id}
         />,