+++ packages/terra-doc-template/CHANGELOG.md
@@ -4,6 +4,15 @@ ChangeLog
 Unreleased
 ----------
 
+1.1.0 - (May 9, 2018)
+------------------
+### Changed
+* Add syntax highlighting to source code examples
+* Add a link to the component's source code
+
+### Fixed
+* Update hash key reference in render to match the defined key in the propsTables prop
+
 1.0.0 - (May 2, 2018)
 ------------------
 * Initial stable release